+++ tests/utils.py
@@ -3,6 +3,7 @@ import pkg_resources
 from pathlib import Path
 import os
 import tempfile
+import yaml
 
 import circuit_build
 
@@ -21,3 +22,23 @@ def tmp_cwd():
             yield tmpdirname
         finally:
             os.chdir(original_cwd)
+
+
+@contextmanager
+def edit_yaml(yaml_file):
+    """Context manager within which you can edit a yaml file.
+
+    Args:
+        yaml_file (Path): path to a yaml file
+
+    Returns:
+        Yields a dict instance of `yaml_file`. This instance will be saved later on the context
+            manager leave.
+    """
+    with yaml_file.open('r') as f:
+        config = yaml.safe_load(f)
+    try:
+        yield config
+    finally:
+        with yaml_file.open('w') as f:
+            yaml.safe_dump(config, f)